[Feature #159994135] Use Django filters in the ManageAssets class
